+++ src/hittable.rs
@@ -8,19 +8,19 @@ use crate::{
 };
 
 pub trait Hittable<'a, T: Rng>: Sync {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>>;
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB>;
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>>;
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB>;
 }
 
 pub struct Hit<'a, T: Rng> {
-    t: f32,
+    t: f64,
     p: Vec3,
     normal: Vec3,
     material: &'a dyn Scatterable<T>,
 }
 
 impl<T: Rng> Hit<'_, T> {
-    pub fn new(t: f32, p: Vec3, normal: Vec3, material: &'_ dyn Scatterable<T>) -> Hit<'_, T> {
+    pub fn new(t: f64, p: Vec3, normal: Vec3, material: &'_ dyn Scatterable<T>) -> Hit<'_, T> {
         Hit {
             t,
             p,
@@ -29,7 +29,7 @@ impl<T: Rng> Hit<'_, T> {
         }
     }
 
-    pub fn t(&self) -> f32 {
+    pub fn t(&self) -> f64 {
         self.t
     }
 
@@ -50,20 +50,20 @@ impl<T: Rng> Hit<'_, T> {
 pub struct Sphere {
     center_initial: Vec3,
     center_final: Vec3,
-    radius: f32,
+    radius: f64,
     material: Material,
-    time_start: f32,
-    time_end: f32,
+    time_start: f64,
+    time_end: f64,
 }
 
 impl Sphere {
     pub fn new(
         center_initial: Vec3,
         center_final: Vec3,
-        radius: f32,
+        radius: f64,
         material: Material,
-        time_start: f32,
-        time_end: f32,
+        time_start: f64,
+        time_end: f64,
     ) -> Sphere {
         Sphere {
             center_initial,
@@ -75,9 +75,9 @@ impl Sphere {
         }
     }
 
-    pub fn center(&self, time: f32) -> Vec3 {
-        let elapsed_time = (time - self.time_start) as f32;
-        let movement_time = (self.time_end - self.time_start) as f32;
+    pub fn center(&self, time: f64) -> Vec3 {
+        let elapsed_time = (time - self.time_start) as f64;
+        let movement_time = (self.time_end - self.time_start) as f64;
 
         let distance = &self.center_final - &self.center_initial;
 
@@ -86,7 +86,7 @@ impl Sphere {
 }
 
 impl<'a, T: Rng> Hittable<'a, T> for Sphere {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>> {
         let oc = ray.origin() - &self.center(ray.time());
 
         let a = ray.direction().dot(ray.direction());
@@ -119,7 +119,7 @@ impl<'a, T: Rng> Hittable<'a, T> for Sphere {
         Some(hit)
     }
 
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB> {
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB> {
         let f = |center: &Vec3| {
             let min = center - Vec3::new(self.radius, self.radius, self.radius);
             let max = center + Vec3::new(self.radius, self.radius, self.radius);
@@ -147,23 +147,23 @@ pub enum Shape {
 }
 
 impl Shape {
-    pub fn sphere(x: f32, y: f32, z: f32, radius: f32, material: Material) -> Shape {
+    pub fn sphere(x: f64, y: f64, z: f64, radius: f64, material: Material) -> Shape {
         let center = Vec3::new(x, y, z);
         let sphere: Sphere = Sphere::new(center.clone(), center, radius, material, 0.0, 1.0);
         Shape::Sphere(sphere)
     }
 
     pub fn moving_sphere(
-        x0: f32,
-        y0: f32,
-        z0: f32,
-        x1: f32,
-        y1: f32,
-        z1: f32,
-        radius: f32,
+        x0: f64,
+        y0: f64,
+        z0: f64,
+        x1: f64,
+        y1: f64,
+        z1: f64,
+        radius: f64,
         material: Material,
-        time_start: f32,
-        time_end: f32,
+        time_start: f64,
+        time_end: f64,
     ) -> Shape {
         let center_initial = Vec3::new(x0, y0, z0);
         let center_final = Vec3::new(x1, y1, z1);
@@ -180,13 +180,13 @@ impl Shape {
 }
 
 impl<'a, T: Rng> Hittable<'a, T> for Shape {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>> {
         match self {
             Shape::Sphere(sphere) => sphere.hit(ray, t_min, t_max),
         }
     }
 
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB> {
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB> {
         match self {
             Shape::Sphere(sphere) => {
                 let sphere: &dyn Hittable<'a, T> = sphere;
@@ -197,37 +197,37 @@ impl<'a, T: Rng> Hittable<'a, T> for Shape {
 }
 
 impl<'a, T: Rng> Hittable<'a, T> for Vec<Shape> {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>> {
         let shapes = &self[..];
         hit(shapes, ray, t_min, t_max)
     }
 
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB> {
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB> {
         bounding_box::<T>(self, time_start, time_end)
     }
 }
 
 impl<'a, T: Rng> Hittable<'a, T> for &'a mut [Shape] {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>> {
         hit(self, ray, t_min, t_max)
     }
 
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB> {
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB> {
         bounding_box::<T>(self, time_start, time_end)
     }
 }
 
 impl<'a, T: Rng> Hittable<'a, T> for &'a [Shape] {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'_, T>> {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'_, T>> {
         hit(self, ray, t_min, t_max)
     }
 
-    fn bounding_box(&self, time_start: f32, time_end: f32) -> Option<AABB> {
+    fn bounding_box(&self, time_start: f64, time_end: f64) -> Option<AABB> {
         bounding_box::<T>(self, time_start, time_end)
     }
 }
 
-fn hit<'a, T: Rng>(shapes: &'a [Shape], ray: &Ray, t_min: f32, t_max: f32) -> Option<Hit<'a, T>> {
+fn hit<'a, T: Rng>(shapes: &'a [Shape], ray: &Ray, t_min: f64, t_max: f64) -> Option<Hit<'a, T>> {
     let mut min_distance = t_max;
     let mut nearest_hit = None;
 
@@ -241,7 +241,7 @@ fn hit<'a, T: Rng>(shapes: &'a [Shape], ray: &Ray, t_min: f32, t_max: f32) -> Op
     nearest_hit
 }
 
-fn bounding_box<'a, T: Rng>(shapes: &[Shape], time_start: f32, time_end: f32) -> Option<AABB> {
+fn bounding_box<'a, T: Rng>(shapes: &[Shape], time_start: f64, time_end: f64) -> Option<AABB> {
     let mut aabb = shapes.first().and_then(|it| {
         let it: &dyn Hittable<'a, T> = it;
         it.bounding_box(time_start, time_end)
