+++ src/ray.rs
@@ -4,11 +4,11 @@ use crate::vec3::Vec3;
 pub struct Ray {
     origin: Vec3,
     direction: Vec3,
-    time: f32,
+    time: f64,
 }
 
 impl Ray {
-    pub fn new(origin: Vec3, direction: Vec3, time: f32) -> Ray {
+    pub fn new(origin: Vec3, direction: Vec3, time: f64) -> Ray {
         Ray {
             origin,
             direction,
@@ -24,11 +24,11 @@ impl Ray {
         &self.direction
     }
 
-    pub fn time(&self) -> f32 {
+    pub fn time(&self) -> f64 {
         self.time
     }
 
-    pub fn point_at(&self, t: f32) -> Vec3 {
+    pub fn point_at(&self, t: f64) -> Vec3 {
         &self.origin + t * &self.direction
     }
 }
