+++ src/texture.rs
@@ -2,7 +2,7 @@ use crate::{perlin::Perlin, vec3::Vec3};
 use rand::Rng;
 
 pub trait Texturable {
-    fn value(&self, u: f32, v: f32, p: &Vec3) -> Vec3;
+    fn value(&self, u: f64, v: f64, p: &Vec3) -> Vec3;
 }
 
 #[derive(Clone, Debug)]
@@ -17,7 +17,7 @@ impl ConstantTexture {
 }
 
 impl Texturable for ConstantTexture {
-    fn value(&self, u: f32, v: f32, p: &Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: &Vec3) -> Vec3 {
         self.color.clone()
     }
 }
@@ -38,7 +38,7 @@ impl CheckerTexture {
 }
 
 impl Texturable for CheckerTexture {
-    fn value(&self, u: f32, v: f32, p: &Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: &Vec3) -> Vec3 {
         let sines = (10.0 * p.x()).sin() * (10.0 * p.y()).sin() * (10.0 * p.z()).sin();
         if sines < 0.0 {
             self.odd.value(u, v, p)
@@ -62,7 +62,7 @@ impl NoiseTexture {
 }
 
 impl Texturable for NoiseTexture {
-    fn value(&self, _u: f32, _v: f32, p: &Vec3) -> Vec3 {
+    fn value(&self, _u: f64, _v: f64, p: &Vec3) -> Vec3 {
         Vec3::new(1.0, 1.0, 1.0) * self.noise.noise(p)
     }
 }
@@ -75,19 +75,19 @@ pub enum Texture {
 }
 
 impl Texture {
-    pub fn constant(r: f32, g: f32, b: f32) -> Self {
+    pub fn constant(r: f64, g: f64, b: f64) -> Self {
         let color = Vec3::new(r, g, b);
         let texture = ConstantTexture::new(color);
         Self::Constant(texture)
     }
 
     pub fn checker_color(
-        r_even: f32,
-        g_even: f32,
-        b_even: f32,
-        r_odd: f32,
-        g_odd: f32,
-        b_odd: f32,
+        r_even: f64,
+        g_even: f64,
+        b_even: f64,
+        r_odd: f64,
+        g_odd: f64,
+        b_odd: f64,
     ) -> Self {
         let even = Self::constant(r_even, g_even, b_even);
         let odd = Self::constant(r_odd, g_odd, b_odd);
@@ -108,7 +108,7 @@ impl Texture {
 }
 
 impl Texturable for Texture {
-    fn value(&self, u: f32, v: f32, p: &Vec3) -> Vec3 {
+    fn value(&self, u: f64, v: f64, p: &Vec3) -> Vec3 {
         match self {
             Texture::Constant(texture) => texture.value(u, v, p),
             Texture::Checker(texture) => texture.value(u, v, p),
