+++ src/vec3.rs
@@ -10,7 +10,7 @@ use std::{
 
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct Vec3 {
-    e: [f32; 3],
+    e: [f64; 3],
 }
 
 impl Vec3 {
@@ -18,48 +18,48 @@ impl Vec3 {
         Vec3::default().unit()
     }
 
-    pub fn new(e0: f32, e1: f32, e3: f32) -> Vec3 {
+    pub fn new(e0: f64, e1: f64, e3: f64) -> Vec3 {
         let e = [e0, e1, e3];
         Vec3 { e }
     }
 
     #[inline(always)]
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.e[0]
     }
 
     #[inline(always)]
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.e[1]
     }
 
     #[inline(always)]
-    pub fn z(&self) -> f32 {
+    pub fn z(&self) -> f64 {
         self.e[2]
     }
 
     #[inline(always)]
-    pub fn r(&self) -> f32 {
+    pub fn r(&self) -> f64 {
         self.x()
     }
 
     #[inline(always)]
-    pub fn g(&self) -> f32 {
+    pub fn g(&self) -> f64 {
         self.y()
     }
 
     #[inline(always)]
-    pub fn b(&self) -> f32 {
+    pub fn b(&self) -> f64 {
         self.z()
     }
 
     #[inline(always)]
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.square_length().sqrt()
     }
 
     #[inline(always)]
-    pub fn square_length(&self) -> f32 {
+    pub fn square_length(&self) -> f64 {
         self.x() * self.x() + self.y() * self.y() + self.z() * self.z()
     }
 
@@ -76,7 +76,7 @@ impl Vec3 {
         self
     }
 
-    pub fn dot(&self, other: &Self) -> f32 {
+    pub fn dot(&self, other: &Self) -> f64 {
         self.x() * other.x() + self.y() * other.y() + self.z() * other.z()
     }
 
@@ -109,9 +109,9 @@ impl<'de> Deserialize<'de> for Vec3 {
             where
                 V: MapAccess<'de>,
             {
-                let mut x: Option<f32> = None;
-                let mut y: Option<f32> = None;
-                let mut z: Option<f32> = None;
+                let mut x: Option<f64> = None;
+                let mut y: Option<f64> = None;
+                let mut z: Option<f64> = None;
 
                 while let Some((key, value)) = map.next_entry()? {
                     match key {
@@ -214,10 +214,10 @@ impl Mul<Vec3> for &Vec3 {
     }
 }
 
-impl Mul<f32> for &Vec3 {
+impl Mul<f64> for &Vec3 {
     type Output = Vec3;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         let x = self.x() * rhs;
         let y = self.y() * rhs;
         let z = self.z() * rhs;
@@ -226,15 +226,15 @@ impl Mul<f32> for &Vec3 {
     }
 }
 
-impl Mul<f32> for Vec3 {
+impl Mul<f64> for Vec3 {
     type Output = Vec3;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         rhs * &self
     }
 }
 
-impl Mul<Vec3> for f32 {
+impl Mul<Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, rhs: Vec3) -> Self::Output {
@@ -242,7 +242,7 @@ impl Mul<Vec3> for f32 {
     }
 }
 
-impl Mul<&Vec3> for f32 {
+impl Mul<&Vec3> for f64 {
     type Output = Vec3;
 
     fn mul(self, rhs: &Vec3) -> Self::Output {
@@ -258,8 +258,8 @@ impl MulAssign for Vec3 {
     }
 }
 
-impl MulAssign<f32> for Vec3 {
-    fn mul_assign(&mut self, rhs: f32) {
+impl MulAssign<f64> for Vec3 {
+    fn mul_assign(&mut self, rhs: f64) {
         self.e[0] *= rhs;
         self.e[1] *= rhs;
         self.e[2] *= rhs;
@@ -278,10 +278,10 @@ impl Div for Vec3 {
     }
 }
 
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
     type Output = Vec3;
 
-    fn div(self, rhs: f32) -> Self::Output {
+    fn div(self, rhs: f64) -> Self::Output {
         let x = self.x() / rhs;
         let y = self.y() / rhs;
         let z = self.z() / rhs;
@@ -290,7 +290,7 @@ impl Div<f32> for Vec3 {
     }
 }
 
-impl Div<Vec3> for f32 {
+impl Div<Vec3> for f64 {
     type Output = Vec3;
 
     fn div(self, rhs: Vec3) -> Self::Output {
@@ -306,8 +306,8 @@ impl DivAssign for Vec3 {
     }
 }
 
-impl DivAssign<f32> for Vec3 {
-    fn div_assign(&mut self, rhs: f32) {
+impl DivAssign<f64> for Vec3 {
+    fn div_assign(&mut self, rhs: f64) {
         self.e[0] /= rhs;
         self.e[1] /= rhs;
         self.e[2] /= rhs;
@@ -370,7 +370,7 @@ impl Neg for Vec3 {
 }
 
 impl Index<usize> for Vec3 {
-    type Output = f32;
+    type Output = f64;
 
     fn index(&self, index: usize) -> &Self::Output {
         &self.e[index]
